Diffuse Map の反映完了 - リファクタリングはできていない - uvの座標系を適当にしているのでモデルのロードの関数にフラグ追加する

+++ directx12-lib/src/ConstantBuffer.cpp
@@ -7,7 +7,7 @@
 /// <param name="src"></param>
 void ConstantBuffer::copy(void* src_constants)
 {
-    uint8_t* pData = nullptr;
+    int16_t* pData = nullptr;
     if (FAILED(this->resource_->Map(0, nullptr, (void**)&pData))) {
         throw std::exception("ConstantBuffer::copy() : resource->Map() Failed.");
     }
